code/requirements.txt
faiss_cpu==1.8.0.post1
matplotlib==3.7.1
numpy==1.24.3
pandas==1.5.3
Pillow==9.4.0
Requests==2.32.3
scikit_learn==1.2.2
scipy==1.10.1
torch==2.0.1+cu118
torchensemble==0.1.9
torchvision==0.15.2+cu118
tqdm==4.65.0
